+++ packages/angular_devkit/build_angular/src/webpack/configs/index.ts
@@ -7,12 +7,8 @@
  */
 
 export * from './analytics';
-export * from './browser';
 export * from './common';
 export * from './dev-server';
-export * from './server';
 export * from './styles';
-export * from './test';
 export * from './typescript';
-export * from './stats';
 export * from './worker';
